requirements.txt
selenium>=4.9.0
undetected-chromedriver>=3.5.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
tenacity>=8.0.0
